[WIP] Make sure that LinalgFlattenVector is run
This change is to experiment with decomposing:

  linalg.depthwise_conv_1d_nwc_wc

into a series of:

  linalg.conv_1d

(1 1D conv per channel dimension). The pass that implements this
transformation "lives" inside MLIR (on a branch).

SUMMARY OF CHANGES:
1. Add LinalgFlattenVector to all lowering pipelines

I wasn't sure _where_ to add to make sure that it's included in our
pipeline of interest, so I added it everywhere.

2. Update the CPUDefault pipeline to include the Vectoriser

1D convolutions don't get vectorised ATM (IREE selects CPUDefault
lowering expert that excludes the vectoriser). This change is just a
hack to include vectorisation in this pipeline.